🏗️ Convert back to old structure
This way of typing is more complex and slightly harder to read, but it allows for shorter names and leaves more opportunity for scaling up.
